+++ internal/stretch/stretch.go
@@ -2,7 +2,6 @@ package stretch
 
 import "math"
 
-// Mode defines the stretch function.
 type Mode int
 
 const (
@@ -13,26 +12,24 @@ const (
 	HistEq
 )
 
-// Apply applies the stretch to normalized input [0..1].
-func Apply(pixels []float64, mode Mode) []float64 {
-	out := make([]float64, len(pixels))
+func Apply(pixels []float32, mode Mode) []float32 {
+	out := make([]float32, len(pixels))
 	switch mode {
 	case Linear:
 		copy(out, pixels)
 	case Log:
 		for i, v := range pixels {
-			out[i] = math.Log1p(9*v) / math.Log1p(9)
+			out[i] = float32(math.Log1p(9*float64(v)) / math.Log1p(9))
 		}
 	case Asinh:
 		for i, v := range pixels {
-			out[i] = math.Asinh(10*v) / math.Asinh(10)
+			out[i] = float32(math.Asinh(10*float64(v)) / math.Asinh(10))
 		}
 	case Sqrt:
 		for i, v := range pixels {
-			out[i] = math.Sqrt(v)
+			out[i] = float32(math.Sqrt(float64(v)))
 		}
 	case HistEq:
-		// Simple cumulative distribution equalization.
 		hist := make([]int, 256)
 		for _, v := range pixels {
 			idx := int(v * 255)
@@ -44,12 +41,12 @@ func Apply(pixels []float64, mode Mode) []float64 {
 			}
 			hist[idx]++
 		}
-		cdf := make([]float64, 256)
+		cdf := make([]float32, 256)
 		total := float64(len(pixels))
 		sum := 0
 		for i, h := range hist {
 			sum += h
-			cdf[i] = float64(sum) / total
+			cdf[i] = float32(float64(sum) / total)
 		}
 		for i, v := range pixels {
 			idx := int(v * 255)
